Add facilities for interop of numeric arrays

+++ src/c_cpp_interop.cpp
@@ -1,14 +1,6 @@
 
 #include "c_cpp_interop.h"
 
-#ifndef STRDUP
-#ifdef _WIN32
-#define STRDUP _strdup
-#else
-#define STRDUP strdup
-#endif
-#endif // !STRDUP
-
 char** to_ansi_char_array(const std::vector<std::string>& charVec)
 {
 	char** result = new char*[charVec.size()];
@@ -37,7 +29,7 @@ double** to_double_ptr_array(const std::vector<std::vector<double>>& mat)
 	double** result = new double*[mat.size()];
 	for (size_t i = 0; i < mat.size(); i++)
 	{
-		result[i] = createDoubleArray(mat[i]);
+		result[i] = to_double_array(mat[i]);
 	}
 	return result;
 }
@@ -49,12 +41,21 @@ void free_double_ptr_array(double** values, int arrayLength)
 	delete[] values;
 }
 
-std::vector<std::string> to_cpp_string_vector(char** names, int size)
+std::vector<std::string> to_cpp_string_vector(char** names, int size, bool cleanup)
 {
 	std::vector<std::string> v(size);
 	for (size_t i = 0; i < size; i++)
 		v[i] = std::string(names[i]);
-	delete_ansi_string_array(names, size);
+	if (cleanup)
+		delete_ansi_string_array(names, size);
 	return v;
+}
+
+std::vector<double> to_cpp_double_vector(double* values, int size, bool cleanup)
+{
+	std::vector<double> data(values, values + size);
+	if (cleanup)
+		delete_array(values);
+	return data;
 }
